Update relation types for `belongs_to` and `has_many` associations
`belongs_to` can be use for one-to-one or one-to-many relations, but the docs omit the one-to-many possibility. Also, because `has_many` is defined on the "many" side of the relation, it technically represent a many-to-one, not a one-to-many.

These changes may seem pedantic, but they are important distinctions for readers familiar with cardinality terminology.